bot/requirements.txt
hyperliquid-python-sdk==0.22.0
eth-account
